Refactor comment creation and retrieval methods
This commit refactors the code in the weddingController.js file to
improve the readability and maintainability of the comment creation and
retrieval methods. The changes include the introduction of a utility
function called 'fdtn' in the formatedDate.js file, which formats the
creation date of comments. This function is now used to format the
'created_at' field of comments in both 'addComment' and 'getAllComments'
methods.

Additionally, two new methods, 'updateComment' and 'deleteComment', are
added to handle updating and deleting comments respectively. These
methods take the comment ID as a parameter and modify or remove the
corresponding comment from the database.

These changes aim to enhance the user experience by providing a
consistent and efficient way to manage comments. They also improve the
code structure by separating concerns and promoting code reuse.